+++ janusgraph-dist/src/test/java/org/janusgraph/pkgtest/CqlAssemblyIT.java
@@ -24,7 +24,7 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 public class CqlAssemblyIT extends AbstractJanusGraphAssemblyIT {
 
     @Container
-    private static JanusGraphCassandraContainer cql = new JanusGraphCassandraContainer(true);
+    private static final JanusGraphCassandraContainer _cql = new JanusGraphCassandraContainer(true);
 
     @Override
     protected String getConfigPath() {