+++ janusgraph-hbase/src/main/java/org/janusgraph/hadoop/config/HBaseAuthHelper.java
@@ -19,7 +19,6 @@ import org.apache.hadoop.mapreduce.Job;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.IOException;
 import java.lang.reflect.Method;
 
 /**
@@ -54,7 +53,7 @@ public class HBaseAuthHelper {
         return inner;
     }
 
-    public static void setHBaseAuthToken(Configuration configuration, Job job) throws IOException {
+    public static void setHBaseAuthToken(Configuration configuration, Job job) {
         // Get HBase authentication token (when configured)
 
         String hbaseAuthentication = configuration.get("hbase.security.authentication");