+++ janusgraph-hbase/src/main/java/org/janusgraph/hadoop/formats/hbase/HBaseBinaryInputFormat.java
@@ -52,12 +52,12 @@ public class HBaseBinaryInputFormat extends AbstractBinaryInputFormat {
     private byte[] edgeStoreFamily;
 
     @Override
-    public List<InputSplit> getSplits(final JobContext jobContext) throws IOException, InterruptedException {
+    public List<InputSplit> getSplits(final JobContext jobContext) throws IOException {
         return this.tableInputFormat.getSplits(jobContext);
     }
 
     @Override
-    public RecordReader<StaticBuffer, Iterable<Entry>> createRecordReader(final InputSplit inputSplit, final TaskAttemptContext taskAttemptContext) throws IOException, InterruptedException {
+    public RecordReader<StaticBuffer, Iterable<Entry>> createRecordReader(final InputSplit inputSplit, final TaskAttemptContext taskAttemptContext) throws IOException {
         tableReader = tableInputFormat.createRecordReader(inputSplit, taskAttemptContext);
         return new HBaseBinaryRecordReader(tableReader, edgeStoreFamily);
     }