+++ janusgraph-hbase/src/main/java/org/janusgraph/hadoop/formats/hbase/HBaseSnapshotBinaryInputFormat.java
@@ -95,7 +95,7 @@ public class HBaseSnapshotBinaryInputFormat extends AbstractBinaryInputFormat {
     }
 
     @Override
-    public RecordReader<StaticBuffer, Iterable<Entry>> createRecordReader(final InputSplit inputSplit, final TaskAttemptContext taskAttemptContext) throws IOException, InterruptedException {
+    public RecordReader<StaticBuffer, Iterable<Entry>> createRecordReader(final InputSplit inputSplit, final TaskAttemptContext taskAttemptContext) throws IOException {
         tableReader = tableSnapshotInputFormat.createRecordReader(inputSplit, taskAttemptContext);
         janusgraphRecordReader = new HBaseBinaryRecordReader(tableReader, edgeStoreFamily);
         return janusgraphRecordReader;
@@ -154,10 +154,5 @@ public class HBaseSnapshotBinaryInputFormat extends AbstractBinaryInputFormat {
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
-    }
-
-    @Override
-    public Configuration getConf() {
-        return super.getConf();
     }
 }