+++ janusgraph-hbase/src/test/java/org/janusgraph/diskstorage/hbase/HBaseStoreManagerConfigTest.java
@@ -79,7 +79,7 @@ public class HBaseStoreManagerConfigTest {
     
     @Test
     // Test HBase preferred timestamp provider MILLI is set by default
-    public void testHBaseTimestampProvider() throws BackendException {
+    public void testHBaseTimestampProvider() {
         // Get an empty configuration
         // GraphDatabaseConfiguration.buildGraphConfiguration() only build an empty one.
         ModifiableConfiguration config = GraphDatabaseConfiguration.buildGraphConfiguration();